# docs/swagger.yaml
basePath: /
definitions:
  handlers.CreateRequest:
    properties:
      end_month:
        type: string
      monthly_price:
        type: string
      service_name:
        type: string
      start_month:
        type: string
      user_id:
        type: string
    type: object
  handlers.SubscriptionDTO:
    properties:
      created_at:
        type: string
      end_month:
        type: string
      id:
        type: string
      monthly_price:
        type: string
      service_name:
        type: string
      start_month:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
    type: object
  handlers.TotalResponse:
    properties:
      from:
        type: string
      service_name:
        type: string
      to:
        type: string
      total:
        type: string
      user_id:
        type: string
    type: object
  handlers.UpdateRequest:
    properties:
      end_month:
        type: string
      monthly_price:
        type: string
      service_name:
        type: string
      start_month:
        type: string
    type: object
info:
  contact: {}
  description: REST API для управления подписками и расчёта сумм за период.
  title: Subscriptions API
  version: "1.0"
paths:
  /subscriptions:
    get:
      parameters:
      - description: Filter by user UUID
        in: query
        name: user_id
        type: string
      - description: Filter by service (ILIKE)
        in: query
        name: service_name
        type: string
      - description: Limit (1..200)
        in: query
        name: limit
        type: integer
      - description: Offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handlers.SubscriptionDTO'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List subscriptions
      tags:
      - subscriptions
    post:
      consumes:
      - application/json
      parameters:
      - description: payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handlers.SubscriptionDTO'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create subscription
      tags:
      - subscriptions
  /subscriptions/{id}:
    delete:
      parameters:
      - description: Subscription ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete subscription
      tags:
      - subscriptions
    get:
      parameters:
      - description: Subscription ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.SubscriptionDTO'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get subscription by id
      tags:
      - subscriptions
    put:
      consumes:
      - application/json
      parameters:
      - description: Subscription ID
        in: path
        name: id
        required: true
        type: string
      - description: payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.UpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.SubscriptionDTO'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update subscription
      tags:
      - subscriptions
  /subscriptions/total:
    get:
      parameters:
      - description: YYYY-MM
        in: query
        name: from
        required: true
        type: string
      - description: YYYY-MM
        in: query
        name: to
        required: true
        type: string
      - description: User UUID
        in: query
        name: user_id
        type: string
      - description: Service filter (ILIKE)
        in: query
        name: service_name
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.TotalResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Total cost for period
      tags:
      - subscriptions
swagger: "2.0"